requirements.txt
oracledb==1.0.2
python-dotenv==0.19.2
unittest2==1.1.0
